refactor(engine): share one lineOffsets helper
